Fixing logit formulation
The initial version of this implemented the relative cost logit the way
it is in the integration branch, which makes it easier to add an
absolute cost logit. However, it was hard to debug since it differed
from GCAM4.3. So, we are switching to the GCAM4.3 equations, but they
are structured the way the integration branch is. This will allow
debugging/testing checks now, but will hopefully still make it easy to
add other logit options later.