Refactor for better Monte Carlo performance.
When we run an Monte Carlo ensemble of scenarios (i.e., parameter
settings), we now store results in a single merged results file,
instead of individual files for each parameter set.  We have also
converted results to .rds files, so that they are smaller on disk, and
we no longer write debugging/diagnostic data during Monte Carlo runs.

As a result of these changes, loading the run data and calculating the
likelihoods and posteriors is a bit more fiddly than it used to be.
The run_bayes function has been updated to handle these changes, and
the subsidiary functions have been removed from the public interface.