Convert ScenarioInfo objects to lists in run_ensemble
The parallel loop potentially serializes the ScenarioInfo objects and
sends them to remote nodes.  It's not entirely clear that it's safe to
do this with an environment because of their internal pointers.
Therefore, to be on the safe side, we first convert the ScenarioInfo
objects to lists and then convert them back to objects in the parallel
workers.