Allow expectation parameters to differ by crop group
Enable differentiation of expectation parameters by crop group. That
is, different crops can use a different number of years for linear
expectation or a different share for the lagged expectation. For
tractability, divide crops into three groups with the option for three
different parameters (as opposed to each crop getting its own
parameter). Set a boolean in `run_ensemble` to determine whether crops
will use the same or different parameters